4/22 한웅재 : 프로젝트 등록, 조회, 상세조회, 수정, 삭제 전부 토큰기반으로 변경 data.sql에 insert company 비밀번호 일부 수정

+++ src/main/java/devconnect/service/ProjectService.java
@@ -24,15 +24,28 @@ public class ProjectService {
     private final ProjectRepository projectRepository;
     private final CompanyRepository companyRepository;
     private final JwtUtil jwtUtil;
+    
+    /// 토큰에 존재하는 데이터를 이용하여 엔티티를 반환하는 함수 <br/>
+    /// 테이블에 값이 없을 경우 null을 리턴
+    public CompanyEntity tokenToCompanyEntity(String token) {
+        String cid = jwtUtil.valnoateToken(token);
+        System.out.println("cid = " + cid);
+        CompanyEntity companyEntity = companyRepository.findByCid(cid);
+        if(companyEntity == null) {
+            System.out.println("companyEntity = null");
+            return null;
+        }
+        System.out.println("companyEntity = " + companyEntity);
+        return companyEntity;
+    }
+    
 
     /// ● 프로젝트 등록
     public boolean writeProject(String token, ProjectDto projectDto) {
         System.out.println("ProjectService.writeProject");
         System.out.println("token = " + token + "\nprojectDto = " + projectDto);
-        String cid = jwtUtil.valnoateToken(token);
-        System.out.println("cid = " + cid);
-        CompanyEntity companyEntity = companyRepository.findByCid(cid);
-        System.out.println("companyEntity = " + companyEntity);
+        // 토큰의 데이터에 있는 회사가 있는지 확인하는 부분
+        CompanyEntity companyEntity = tokenToCompanyEntity(token);
         if(companyEntity != null) {
             projectDto.setCno(companyEntity.getCno());
             System.out.println("projectDto = " + projectDto);
@@ -62,38 +75,63 @@ public class ProjectService {
         return projectDtoList;
     }
 
+    /// ● 프로젝트 상세조회
+    public ProjectDto findProject(String token, int pno) {
+        System.out.println("ProjectService.findProject");
+        System.out.println("pno = " + pno + ", token = \n" + token);
+        // 토큰의 데이터에 있는 회사가 있는지 확인하는 부분
+        CompanyEntity companyEntity = tokenToCompanyEntity(token);
+        if(companyEntity == null) { return null; }
+        Optional<ProjectEntity> optional = projectRepository.findById(pno);
+        if(optional.isPresent()) {
+            ProjectDto projectDto = optional.get().toDto();
+            System.out.println("projectDto = " + projectDto);
+            return projectDto;
+        }
+        return null;
+    }
+
     /// ● 프로젝트 수정
-    // {"pno" : 1, "pname" : "수정 테스트 [백엔드]", "pintro" : "수정 테스트 소개", "ptype" : 1, "pcomment" : "수정 테스트 상세 설명", "pcount" : 2, "pstart" : "2025-04-21T13:00:00", "pend" : "2025-05-20T13:00:00", "recruit_pstart" : "2025-06-15T13:00:00", "recruit_pend" : "2025-12-12T13:00:00", "ppay" : 3200}
-    public boolean updateProject(ProjectDto projectDto) {
+    public boolean updateProject(String token, ProjectDto projectDto) {
         System.out.println("ProjectService.updateProject");
-        System.out.println("projectDto = " + projectDto);
+        System.out.println("token = \n" + token + "\nprojectDto = " + projectDto);
+        // 토큰의 데이터에 있는 회사가 있는지 확인하는 부분
+        CompanyEntity companyEntity = tokenToCompanyEntity(token);
+        if(companyEntity == null) { return false; }
         Optional<ProjectEntity> optional = projectRepository.findById(projectDto.getPno());
         if(optional.isPresent()) {
             ProjectEntity projectEntity = optional.get();
-            if(!projectEntity.getPname().equals(projectDto.getPname())) { projectEntity.setPname(projectDto.getPname()); }
-            if(!projectEntity.getPintro().equals(projectDto.getPintro())) { projectEntity.setPintro(projectDto.getPintro()); }
-            if(!projectEntity.getPcomment().equals(projectDto.getPcomment())) { projectEntity.setPcomment(projectDto.getPcomment()); }
-            if(!projectEntity.getPstart().equals(projectDto.getPstart())) { projectEntity.setPstart(projectDto.getPstart()); }
-            if(!projectEntity.getPend().equals(projectDto.getPend())) { projectEntity.setPend(projectDto.getPend()); }
-            if(!projectEntity.getRecruit_pstart().equals(projectDto.getRecruit_pstart())) { projectEntity.setRecruit_pstart(projectDto.getRecruit_pstart()); }
-            if(!projectEntity.getRecruit_pend().equals(projectDto.getRecruit_pend())) { projectEntity.setRecruit_pend(projectDto.getRecruit_pend()); }
-            if(projectEntity.getPtype() != projectDto.getPtype()) { projectEntity.setPtype(projectDto.getPtype()); }
-            if(projectEntity.getPcount() != projectDto.getPcount()) { projectEntity.setPcount(projectDto.getPcount()); }
-            if(projectEntity.getPpay() != projectDto.getPpay()) { projectEntity.setPpay(projectDto.getPpay()); }
-            return true;
+            if(companyEntity.getCno() == projectEntity.getCompanyEntity().getCno()) {
+                if(!projectEntity.getPname().equals(projectDto.getPname())) { projectEntity.setPname(projectDto.getPname()); }
+                if(!projectEntity.getPintro().equals(projectDto.getPintro())) { projectEntity.setPintro(projectDto.getPintro()); }
+                if(!projectEntity.getPcomment().equals(projectDto.getPcomment())) { projectEntity.setPcomment(projectDto.getPcomment()); }
+                if(!projectEntity.getPstart().equals(projectDto.getPstart())) { projectEntity.setPstart(projectDto.getPstart()); }
+                if(!projectEntity.getPend().equals(projectDto.getPend())) { projectEntity.setPend(projectDto.getPend()); }
+                if(!projectEntity.getRecruit_pstart().equals(projectDto.getRecruit_pstart())) { projectEntity.setRecruit_pstart(projectDto.getRecruit_pstart()); }
+                if(!projectEntity.getRecruit_pend().equals(projectDto.getRecruit_pend())) { projectEntity.setRecruit_pend(projectDto.getRecruit_pend()); }
+                if(projectEntity.getPtype() != projectDto.getPtype()) { projectEntity.setPtype(projectDto.getPtype()); }
+                if(projectEntity.getPcount() != projectDto.getPcount()) { projectEntity.setPcount(projectDto.getPcount()); }
+                if(projectEntity.getPpay() != projectDto.getPpay()) { projectEntity.setPpay(projectDto.getPpay()); }
+                return true;
+            }
         }
         return false;
     }
 
     /// ● 프로젝트 삭제
-    public boolean deleteProject(int pno) {
+    public boolean deleteProject(String token, int pno) {
         System.out.println("ProjectService.deleteProject");
-        System.out.println("pno = " + pno);
+        System.out.println("token = \n" + token + "\npno = " + pno);
+        // 토큰의 데이터에 있는 회사가 있는지 확인하는 부분
+        CompanyEntity companyEntity = tokenToCompanyEntity(token);
+        if(companyEntity == null) { return false; }
         Optional<ProjectEntity> optional = projectRepository.findById(pno);
         if(optional.isPresent()) {
             ProjectEntity projectEntity = optional.get();
-            projectRepository.deleteById(pno);
-            return true;
+            if(companyEntity.getCno() == projectEntity.getCompanyEntity().getCno()) {
+                projectRepository.deleteById(pno);
+                return true;
+            }
         }
         return false;
     }
